requirements.txt
einops==0.8.0
einops_exts==0.0.3
nltk==3.7
Pillow==11.0.0
psutil==5.9.4
scipy==1.14.1
timm==0.6.12
torchtext==0.9.1
tqdm